refactor: Consolidate plugin, queue, and scanner module APIs with strict encapsulation
  - Create api.rs modules as single public interfaces for plugin, queue, and scanner systems
  - Move all module tests into organized tests/ subdirectories
  - Extract shared types into dedicated types.rs and traits.rs files
  - Update all external imports to use controlled api:: interfaces
  - Enforce pub(crate) visibility for internal modules
  - Add missing public methods (scanner_count) for test compatibility
  - Reorganize queue tests from individual files into tests/ structure
  - Extract plugin error tests into dedicated test module
  - Move scanner manager_tests.rs into organized tests/manager.rs structure
  - Update notification and core service imports for new API boundaries
