+++ src/notifications/tests/unit/subscriber_tests.rs
@@ -1,7 +1,8 @@
 //! Unit tests for Subscriber trait and SubscriberStatistics
 
-use crate::notifications::event::{Event, SystemEvent, SystemEventType};
-use crate::notifications::traits::{Subscriber, SubscriberStatistics};
+use crate::notifications::api::{
+    Event, Subscriber, SubscriberStatistics, SystemEvent, SystemEventType,
+};
 use async_trait::async_trait;
 use std::time::Instant;
