+++ src/plugin/discovery.rs
@@ -1,10 +1,7 @@
-//! Plugin Discovery System (Legacy - Removed)
+//! Plugin discovery system
 //!
-//! This file previously contained the legacy plugin discovery system which has been
-//! replaced by the unified discovery system in unified_discovery.rs.
-//!
-//! The legacy system focused on trait-based discovery with PluginFactory enums,
-//! while the new system uses struct-based discovery with PluginSource enums
-//! to better support external cdylib plugin loading.
+//! This module provides plugin discovery functionality, delegating to the
+//! unified discovery system while maintaining API consistency.
 
-// This module is now empty - all functionality moved to unified_discovery.rs
+pub use crate::plugin::types::{DiscoveredPlugin, PluginSource};
+pub use crate::plugin::unified_discovery::PluginDiscovery;
