+++ src/plugin/mod.rs
@@ -3,25 +3,21 @@
 //! Provides a trait-based interface for plugin communication with async notifications.
 //! Supports dynamic plugin loading, version compatibility, and real-time notifications.
 
-pub mod args;
-pub mod builtin;
-pub mod context;
-pub mod discovery;
-pub mod error;
-pub mod manager;
-pub mod registry;
-pub mod settings;
-pub mod traits;
-pub mod unified_discovery;
+// Internal modules - all access should go through api module
+pub(crate) mod args;
+pub(crate) mod builtin;
+pub(crate) mod context;
+pub(crate) mod discovery;
+pub(crate) mod error;
+pub(crate) mod manager;
+pub(crate) mod registry;
+pub(crate) mod settings;
+pub(crate) mod traits;
+pub(crate) mod types;
+pub(crate) mod unified_discovery;
 
-// Re-export core types for easier access
-pub use args::{OutputFormat, PluginArgParser, PluginConfig};
-pub use context::PluginContext;
-pub use error::{PluginError, PluginResult};
-pub use manager::{PluginManager, PluginMetadata, PluginProxy};
-pub use registry::{PluginRegistry, SharedPluginRegistry};
-pub use settings::PluginSettings;
-pub use traits::{
-    ConsumerPlugin, Plugin, PluginDataRequirements, PluginFunction, PluginInfo, PluginType,
-};
-pub use unified_discovery::{DiscoveredPlugin, PluginSource};
+// Public API module - the only public interface for the plugin system
+pub mod api;
+
+#[cfg(test)]
+mod tests;
