+++ src/plugin/tests/mod.rs
@@ -1,0 +1,6 @@
+//! Test modules for the plugin system
+//!
+//! This module organizes all the test suites for the plugin system,
+//! including manager tests, registry tests, and integration tests.
+
+mod error;
