+++ src/queue/mod.rs
@@ -73,74 +73,18 @@
 //! # }
 //! ```
 
-mod consumer;
-mod error;
-mod internal;
-mod manager;
-mod message;
-mod publisher;
+// Internal modules - all access should go through api module
+pub(crate) mod consumer;
+pub(crate) mod error;
+pub(crate) mod internal;
+pub(crate) mod manager;
+pub(crate) mod message;
+pub(crate) mod publisher;
+pub(crate) mod traits;
+pub(crate) mod types;
 
-pub use consumer::QueueConsumer;
-pub use error::QueueError;
-pub use internal::MultiConsumerQueue;
-pub use manager::QueueManager;
-pub use message::{GroupedMessage, Message, MessageHeader};
-pub use publisher::QueuePublisher;
+// Public API module - the only public interface for the queue system
+pub mod api;
 
-/// Memory usage statistics for the queue system
-#[derive(Debug, Clone, PartialEq)]
-pub struct MemoryStats {
-    /// Total number of messages in the queue
-    pub total_messages: usize,
-    /// Total memory usage in bytes
-    pub total_bytes: usize,
-    /// Memory used by message data
-    pub message_data_bytes: usize,
-    /// Memory used by Arc and metadata overhead
-    pub overhead_bytes: usize,
-}
-
-/// Consumer lag statistics for the queue system
-#[derive(Debug, Clone, PartialEq)]
-pub struct LagStats {
-    /// Total number of active consumers
-    pub total_consumers: usize,
-    /// Maximum lag among all consumers
-    pub max_lag: usize,
-    /// Minimum lag among all consumers
-    pub min_lag: usize,
-    /// Average lag across all consumers
-    pub avg_lag: f64,
-}
-
-/// Information about a stale consumer
-#[derive(Debug, Clone)]
-pub struct StaleConsumerInfo {
-    /// Internal consumer ID
-    pub consumer_id: u64,
-    /// Current lag in messages
-    pub lag: usize,
-    /// Time since last read (seconds)
-    pub seconds_since_last_read: u64,
-}
-
-pub type QueueResult<T> = Result<T, QueueError>;
-
-#[cfg(test)]
-mod concurrent_tests;
-#[cfg(test)]
-mod consumer_tests;
-#[cfg(test)]
-mod edge_case_tests;
-#[cfg(test)]
-mod grouping_tests;
-#[cfg(test)]
-mod integration_tests;
-#[cfg(test)]
-mod lifecycle_tests;
-#[cfg(test)]
-mod memory_tests;
-#[cfg(test)]
-mod publisher_tests;
 #[cfg(test)]
 mod tests;
