+++ src/queue/tests/consumer.rs
@@ -2,7 +2,7 @@
 
 #[cfg(test)]
 mod tests {
-    use crate::queue::{Message, QueueManager};
+    use crate::queue::api::{Message, QueueManager};
     use std::sync::Arc;
 
     #[test]
