+++ src/queue/tests/mod.rs
@@ -1,0 +1,14 @@
+//! Test modules for the queue system
+//!
+//! This module organizes all the test suites for the multiconsumer queue system.
+//! Tests are organized by functional area for better maintainability.
+
+mod concurrent;
+mod consumer;
+mod edge_cases;
+mod grouping;
+mod integration;
+mod lifecycle;
+mod manager;
+mod memory;
+mod publisher;
