+++ src/scanner/task/mod.rs
@@ -9,9 +9,3 @@ mod git_ops;
 mod queue_ops;
 
 pub use core::ScannerTask;
-
-// Re-export for backwards compatibility if needed
-pub use crate::scanner::types::{
-    ChangeType, CommitInfo, FileChangeData, RepositoryData, RepositoryDataBuilder, ScanMessage,
-    ScanStats,
-};
