+++ src/scanner/tests/mod.rs
@@ -1,0 +1,7 @@
+//! Test modules for the scanner system
+//!
+//! This module organizes all the test suites for the scanner system,
+//! including manager tests and types tests.
+
+pub mod manager;
+pub mod types;
